Put Bricolage in root layout to remove hydration
