+++ app/RootLayout.js
@@ -10,6 +10,14 @@ import Footer from "../components/Footer";
 import { FormProvider } from "@/context/FormContext";
 import { AuthorProvider } from "@/context/AuthorContext";
 
+import { Bricolage_Grotesque } from "next/font/google";
+const bricolage = Bricolage_Grotesque({
+  subsets: ["latin"],
+  weight: ["400", "500", "600", "700"],
+  style: ["normal"],
+  display: "swap",
+});
+
 config.autoAddCss = false;
 
 export default function RootLayout({ children }) {
@@ -63,7 +71,10 @@ export default function RootLayout({ children }) {
             <div>
               {/* <AnimatedLayout> */}
               <div className="">
-                <TopNav styles={shouldHideHeader} />
+                <TopNav
+                  styles={shouldHideHeader}
+                  fontClass={bricolage.className}
+                />
                 <div>{children}</div>
                 <Footer styles={shouldHideHeader} />
                 {/* <ScrollToTop styles={hideIfAuthorPage} /> */}
